Separete file for RedisQueue and more comments
